+++ variables-naming.tf
@@ -1,12 +1,12 @@
 # Generic naming variables
 variable "name_prefix" {
-  description = "Optional prefix for the generated name"
+  description = "Optional prefix for the generated name."
   type        = string
   default     = ""
 }
 
 variable "name_suffix" {
-  description = "Optional suffix for the generated name"
+  description = "Optional suffix for the generated name."
   type        = string
   default     = ""
 }
@@ -19,25 +19,25 @@ variable "use_caf_naming" {
 
 # Custom naming override
 variable "custom_vmss_name" {
-  description = "Custom name for the Virtual Machine ScaleSet"
+  description = "Custom name for the Virtual Machine ScaleSet."
   type        = string
   default     = null
 }
 
 variable "custom_nic_name" {
-  description = "Custom name for Network Interfaces"
+  description = "Custom name for Network Interfaces."
   type        = string
   default     = null
 }
 
 variable "custom_ipconfig_name" {
-  description = "Custom name for Ipconfiguration"
+  description = "Custom name for Ipconfiguration."
   type        = string
   default     = null
 }
 
 variable "custom_dcr_name" {
-  description = "Custom name for Data collection rule association"
+  description = "Custom name for Data collection rule association."
   type        = string
   default     = null
 }