+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+Added
+  * AZ-985: Add option to disable Azure Monitor: Agent + Data Collection Rule
+
+Changed
+  * AZ-985: Update Azure Monitor resources naming to align with new condition
+
 # v7.1.1 - 2023-01-13
 
 Fixed